fix comments: rename file SophiaUtil -> sophia-util
